requirements.txt
python-ldap
easygui
pycups
selenium
requests
bottle
